+++ src/platform/mod.rs
@@ -339,7 +339,7 @@ pub unsafe fn init(auxvs: Box<[[usize; 2]]>) {
     redox_rt::sys::this_proc_call(
         &mut [],
         syscall::CallFlags::empty(),
-        &[redox_rt::protocol::ProcCall::SyncSigPctl as usize],
+        &[redox_rt::protocol::ProcCall::SyncSigPctl as u64],
     )
     .expect("failed to sync signal pctl");
